requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
urllib3==2.1.0
python-dotenv==1.0.0
openai==1.54.0
httpx==0.27.0
gunicorn==21.2.0